+++ app/routes/index.tsx
@@ -1,28 +1,8 @@
-import mainIcon from "../assets/blacksheep_100x100.webp";
 import { getAllPostFrontmatter } from "~/utils/blogPosts";
 import { Link, useLoaderData } from "@remix-run/react";
 import type { FrontMatterPlusSlug } from "utils/frontmatterTypings";
-import { DatePublished } from "~/components/DatePublished/DatePublished";
 import { SheepImage } from "~/components/SheepImage/SheepImage";
-
-
-
-export function ListOfArticles(props: {
-  allFrontmatter: Array<FrontMatterPlusSlug>
-}) {
-  return <div data-testid="bsc-homepage-all-article-previews">
-    {props.allFrontmatter.map((v) => {
-      return <Link to={v.slug} key={v.slug} className="homepage-article-link" data-testid="bsc-homepage-article-preview-box">
-        <div>
-          <p className="homepage-article-link-title">{v.frontmatter.meta?.title ?? v.slug}</p>
-          {v.frontmatter.meta?.dateCreated && <DatePublished date={v.frontmatter.meta.dateCreated} />}
-          <p className="homepage-article-link-description">{v.frontmatter.meta?.description ?? ''}</p>
-        </div>
-      </Link>
-    })}
-
-  </div>
-}
+import { ListOfArticles } from "~/components/ListOfArticles/ListOfArticles";
 
 
 export async function loader() {
